+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - xunit.runner.utility 2.4.1
 - xunit.runner.devices 2.5.25
 - fix certificate pin in tests
+- Karamunting.Android.Square.OkHttp 3.14.0
+- Karamunting.Android.Square.Okio 1.17.3
+- Karamunting.Square.OkHttp3.UrlConnection 3.14.0
+- min Android version 21 (for okhttp 3.13.0)
 
 ## 1.11.0
 - xamarin : 15.8.5 (xamarin.vs 4.11.0.776 ; xamarin.android 9.0.0.19 ; xamarin.ios 12.0.0.15 ; vs for mac 7.6.8.38)